fix(#1692): keep refresh path Send + campsite split
- Make schema-discovery infer_column_type's boxed future `+ Send` so the
  registry get_schema future (auto-discovery on miss) is Send. My eager
  resolve_registry_schema is awaited on the napi refresh() path, which
  requires Send; without this the node bindings failed to compile (E0277).
- Move resolve_registry_schema out of the 1184-line reader/mod.rs into a
  dedicated reader/registry_schema.rs (epic #1116 doctrine) so the
  over-threshold mod.rs does not grow materially; residual +5 is the
  cfg-gated field init + mod declaration (unavoidable; full mod.rs split is
  out of scope for a block_on fix).
- cargo fmt.
